Rename project to pyetl-framework
There were several variations of the name floating around.
pyetl-framework better describes the problem that this project is
attempting to solve.
