+++ .env
@@ -1,3 +1,3 @@
 source env/bin/activate
 export APP_SETTINGS='DevelopmentConfig'
-export APP_BASEDIR=$(pwd)/pyscraper_framework
+export APP_BASEDIR=$(pwd)/pyetl_framework
