+++ pyetl_framework/lib/extractor.py
