+++ pyetl_framework/lib/pipeline_manager.py
