+++ pyetl_framework/templates/index.jade
@@ -1,7 +1,7 @@
 extends ./layout.jade
 
 block title
-  |pyscraper - Home
+  |pyetl - Home
 
 block content
   .row
